🎨 account 폴더 생성 & 컴포넌트 파일 분리
