+++ src/components/myprofile/account/NicknameSection.tsx
@@ -1,0 +1,114 @@
+import TextInput from "@/components/common/TextInput";
+import ClientApi from "@/lib/clientApi";
+import { CircleAlert } from "lucide-react";
+import { useEffect, useOptimistic, useState, useTransition } from "react";
+
+interface NicknameProps {
+  initialNickname: string;
+}
+
+export default function NicknameSection({ initialNickname }: NicknameProps) {
+  const [nickname, setNickname] = useState<string>("");
+  const [modifyNickname, setModifyNickname] = useState<string>("");
+  const [isNicknameEditing, setIsNicknameEditing] = useState<boolean>(false);
+  const [nicknameError, setNicknameError] = useState<string | null>(null);
+
+  const [isNicknamePending, startNicknameTransition] = useTransition();
+  const [optimisticNickname, addOptimisticNickname] = useOptimistic<
+    string | null,
+    string
+  >(nickname, (_: string | null, nextValue: string) => nextValue);
+
+  useEffect(() => {
+    setNickname(initialNickname);
+  }, [initialNickname]);
+
+  const handleNicknameSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setNicknameError(null);
+    if (isNicknamePending) return;
+    const parsedNickname = nicknameSchema.safeParse(modifyNickname);
+
+    if (!parsedNickname.success) {
+      setNicknameError(parsedNickname.error.issues[0]?.message);
+      return;
+    }
+
+    startNicknameTransition(async () => {
+      addOptimisticNickname(parsedNickname.data);
+      try {
+        const res = await ClientApi("/api/v1/users/me/nickname", {
+          method: "PATCH",
+          body: JSON.stringify({ nickname: parsedNickname.data }),
+        });
+
+        if (res.ok) {
+          setNickname(modifyNickname);
+          setIsNicknameEditing(false);
+        } else {
+          alert("닉네임 변경에 실패했습니다.");
+        }
+      } catch (error) {
+        alert("서버 통신 중 오류가 발생했습니다.");
+      }
+    });
+  };
+  return (
+    <div className="space-y-4">
+      <h3>닉네임</h3>
+      <div className="space-y-2">
+        {isNicknameEditing ? (
+          <form
+            className="flex flex-row items-center space-x-4"
+            onSubmit={handleNicknameSubmit}
+          >
+            <TextInput
+              value={modifyNickname}
+              onChange={(e) => setModifyNickname(e.target.value)}
+              placeholder={nickname ?? ""}
+              className="h-5 w-42 py-4"
+            />
+
+            <button
+              type="submit"
+              className="text-accent cursor-pointer hover:underline"
+            >
+              {isNicknamePending ? "저장 중..." : "저장"}
+            </button>
+            <button
+              type="button"
+              onClick={() => {
+                setIsNicknameEditing(false);
+                setNicknameError(null);
+              }}
+              className="text-content-secondary cursor-pointer hover:underline"
+            >
+              취소
+            </button>
+          </form>
+        ) : (
+          <div className="flex flex-row space-x-4">
+            <span>{optimisticNickname}</span>
+            <button
+              className="text-accent cursor-pointer hover:underline"
+              onClick={() => {
+                console.log("optimisticNickname", optimisticNickname);
+                setModifyNickname(nickname ?? "");
+                setIsNicknameEditing(true);
+              }}
+            >
+              수정
+            </button>
+          </div>
+        )}
+        {nicknameError && (
+          <p className="text-negative ml-2 text-sm">{nicknameError}</p>
+        )}
+        <div className="text-content-secondary flex items-center gap-2 text-base">
+          <CircleAlert size={18} />
+          <p>닉네임은 7일 단위로 변경할 수 있습니다</p>
+        </div>
+      </div>
+    </div>
+  );
+}
