+++ src/components/myprofile/account/PasswordSection.tsx
@@ -1,0 +1,72 @@
+import { BoxButton } from "@/components/common/button/BoxButton";
+import TextInput from "@/components/common/TextInput";
+import ClientApi from "@/lib/clientApi";
+import { useState } from "react";
+
+export default function PasswordSection() {
+  const [password, setPassword] = useState<string>("");
+  const [newPassword, setNewPassword] = useState<string>("");
+  const [newPasswordConfirm, setNewPasswordConfirm] = useState<string>("");
+  const [passwordError, setPasswordError] = useState<string | null>(null);
+
+  const handlePasswordSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (newPassword !== newPasswordConfirm) {
+      alert("새 비밀번호와 새 비밀번호 확인이 일치하지 않습니다.");
+      setPasswordError("새 비밀번호와 새 비밀번호 확인이 일치하지 않습니다.");
+      return;
+    }
+    const parsedNewPW = passwordSchema.safeParse(newPassword);
+    if (!parsedNewPW.success) {
+      setPasswordError(parsedNewPW.error.issues[0]?.message);
+      return;
+    }
+
+    try {
+      const res = await ClientApi("/api/v1/users/me/password", {
+        method: "PATCH",
+        body: JSON.stringify({
+          password: password,
+          newPassword: newPassword,
+          newPasswordConfirm: newPasswordConfirm,
+        }),
+      });
+
+      if (res.ok) {
+        alert("비밀번호 변경 성공!");
+      } else {
+        alert("비밀번호 변경에 실패했습니다.");
+      }
+    } catch (error) {
+      alert("서버 통신 중 오류가 발생했습니다.");
+    }
+  };
+
+  return (
+    <div className="flex flex-col gap-2">
+      <h3>비밀번호 변경</h3>
+      <form className="flex flex-col gap-2">
+        <TextInput
+          placeholder="현재 비밀번호"
+          onChange={(e) => setPassword(e.target.value)}
+        />
+        <TextInput
+          placeholder="새 비밀번호"
+          onChange={(e) => setNewPassword(e.target.value)}
+        />
+        <TextInput
+          placeholder="새 비밀번호 확인"
+          onChange={(e) => setNewPasswordConfirm(e.target.value)}
+        />
+        <span className="ml-2">{passwordError}</span>
+        <BoxButton
+          text="수정"
+          tone="color"
+          size="xs"
+          className="mt-3 flex-nowrap self-end"
+          onClick={handlePasswordSubmit}
+        />
+      </form>
+    </div>
+  );
+}
